update terminal changelog format
