+++ projects/openapi-utilities/src/utilities/changelog-renderers/__tests__/json-changelog.test.ts
@@ -1,9 +1,9 @@
 import { test, expect } from '@jest/globals';
-import { openAPI as petStoreBase } from '../../examples/petstore-base';
-import { openAPI as petStoreUpdated } from '../../examples/petstore-updated';
+import { openAPI as petStoreBase } from '../../../examples/petstore-base';
+import { openAPI as petStoreUpdated } from '../../../examples/petstore-updated';
 import { jsonChangelog } from '../json-changelog';
-import { groupDiffsByEndpoint } from '../../openapi3/group-diff';
-import { diff } from '../../diff/diff';
+import { groupDiffsByEndpoint } from '../../../openapi3/group-diff';
+import { diff } from '../../../diff/diff';
 
 test('json changelog collects changes properly', () => {
   const diffs = diff(petStoreBase, petStoreUpdated);
