+++ projects/openapi-utilities/src/utilities/changelog-renderers/json-changelog.ts
@@ -1,15 +1,16 @@
 import { OpenAPIV3 } from 'openapi-types';
 import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
 
-import { typeofDiff, getTypeofDiffs } from '../diff/diff';
+import { typeofDiff, getTypeofDiffs } from '../../diff/diff';
 import type {
   GroupedDiffs,
   Body,
   Diff,
   Endpoint,
   Response,
-} from '../openapi3/group-diff';
-import { getRaw } from '../openapi3/traverser';
+} from '../../openapi3/group-diff';
+import { getRaw } from '../../openapi3/traverser';
+import { interpretFieldLevelDiffs } from './common';
 
 type RawChange<T> = { key: string } & (
   | {
@@ -200,41 +201,12 @@ function getRequestChangeLogs(
   };
 }
 
-function getRootBodyPath(path: string): string {
-  const parts = jsonPointerHelpers.decode(path);
-  if (parts[3] === 'responses') {
-    return jsonPointerHelpers.compile(parts.slice(0, 7));
-  } else {
-    return jsonPointerHelpers.compile(parts.slice(0, 6));
-  }
-}
-
 function getBodyChangeLogs(
   specs: { from: OpenAPIV3.Document; to: OpenAPIV3.Document },
   body: Body,
   contentType: string
 ): ChangedNode {
-  const fieldDiffs: Diff[] = Object.entries(body.fields).map(([key, diffs]) => {
-    const firstDiffPath = diffs[0].after ?? diffs[0].before;
-
-    // TODO figure out if we need to handle path reconciliation here
-    // This might only be an issue with nested enums + rearranging + removing keys with oneOf/anyOf/allOf
-    const absolutePath = jsonPointerHelpers.join(
-      getRootBodyPath(firstDiffPath),
-      key
-    );
-    const beforeRaw = jsonPointerHelpers.tryGet(specs.from, absolutePath);
-    const afterRaw = jsonPointerHelpers.tryGet(specs.to, absolutePath);
-    const before = beforeRaw.match ? absolutePath : undefined;
-    const after = afterRaw.match ? absolutePath : undefined;
-
-    return {
-      before,
-      after,
-      trail: key,
-      change: before && after ? 'changed' : before ? 'removed' : 'added',
-    } as Diff;
-  });
+  const fieldDiffs = interpretFieldLevelDiffs(specs, body.fields);
   const exampleDiffs = body.examples;
 
   // Group body diffs by trail and then log based on that
